feature: Move to wvlet.ai.core
