+++ ai-core/.js/src/main/scala/wvlet/ai/core/log/LogTimestampFormatter.scala
@@ -1,4 +1,4 @@
-package wvlet.ai.log
+package wvlet.ai.core.log
 
 import scala.scalajs.js
