+++ ai-core/.jvm/src/main/scala/wvlet/ai/core/log/AsyncHandler.scala
@@ -1,4 +1,4 @@
-package wvlet.ai.log
+package wvlet.ai.core.log
 
 import wvlet.ai.util.control.{Guard, ThreadUtil}
