+++ ai-core/src/main/scala/wvlet/ai/core/log/LogEnvBase.scala
@@ -1,4 +1,4 @@
-package wvlet.ai.log
+package wvlet.ai.core.log
 
 import java.io.PrintStream
