+++ ai-core/src/main/scala/wvlet/ai/design/Design.scala
@@ -13,10 +13,10 @@
  */
 package wvlet.ai.design
 
+import wvlet.ai.core.log.LogSupport
 import wvlet.ai.design.Binder.Binding
 import wvlet.ai.design.DesignOptions.*
 import wvlet.ai.design.LifeCycleHookType
-import wvlet.ai.log.LogSupport
 import wvlet.ai.surface.Surface
 import wvlet.ai.util.SourceCode
