+++ ai-core/src/main/scala/wvlet/ai/design/DesignStats.scala
@@ -13,8 +13,8 @@
  */
 package wvlet.ai.design
 
+import wvlet.ai.core.log.LogSupport
 import wvlet.ai.design.{Design, Session}
-import wvlet.ai.log.LogSupport
 import wvlet.ai.surface.Surface
 
 import java.util.concurrent.ConcurrentHashMap
